requirements.txt
flask
torch
torchvision
numpy
pillow
werkzeug
gunicorn